Sistemati errori e bug all'interno del codice

+++ lib/cobot/include/cobot.h
@@ -5,8 +5,8 @@
 #include "box.h"
 #include "global_variables.h"
 
-void get_piece_to_box(const char ID_cobot, float conveyor_lenght, float conveyor_speed, system_clock::time_point start_time);
+void get_piece_to_box(const char ID_cobot, uint conveyor_lenght, uint conveyor_speed, system_clock::time_point start_time);
 
 // Funzione per il calcolo del tempo di simulazione del cobot:
-float calc_sleep_time(float conveyor_lenght, float conveyor_speed);
+uint calc_sleep_time(uint conveyor_lenght, uint conveyor_speed);
 #endif
